Moved sm_utils.hpp to utils.
