+++ cpp/include/raft/utils/sm_utils.hpp
